First step to superclass Core
plus fix in deleteNode
